Add files via upload
The project dealt with many concepts related to System for electronic medical record , When using this system within a hospital, It is observed that medical records are being kept in physical folders and then placed on cabinets, It was also noted from the record section that patients (students and staff) records are stored in many formats and these results to disorganizations of the records, leading to a higher risk of medical errors, duplicate procedures, and time loss in searching and obtaining information Although the paper records system provides accountability, it has inherent challenges for example difficult access, takes a long time to update, is not secured and it is impossible to share quickly between different sites and keep it for too long without damaging it may lead to permanent loss For valuable patient data their indirect effect may be impossible to quantify. A fully nationalized electronic health records system can address and improve some of these problems.